requirements.txt
streamlit
torch
torchvision
Pillow
json5
